E2E: Revamp `Reader` and `Notifications` spec to reduce overlap. (#59162)
* Revamp multiple specs centred around notifications.
- eliminate commenting steps in the Reader spec and move it inside the Notificatiosn spec.
- clear the authentication state instead of launching a new context to trash the comment.

* Fix eslint style

* Rebase onto trunk.
