+++ test/e2e/specs/specs-playwright/wp-notifications__trash.ts
@@ -4,8 +4,7 @@
 
 import {
 	DataHelper,
-	PublishedPostsListPage,
-	CommentsComponent,
+	ReaderPage,
 	NavbarComponent,
 	NotificationsComponent,
 	TestAccount,
@@ -19,41 +18,34 @@ describe( DataHelper.createSuiteTitle( 'Notifications' ), function () {
 	const notificationsUser = 'notificationsUser';
 	const comment = DataHelper.getRandomPhrase() + ' notifications-trash-spec';
 
-	let page: Page;
-	let publishedPostsListPage: PublishedPostsListPage;
 	let notificationsComponent: NotificationsComponent;
-
-	beforeAll( async () => {
-		page = await browser.newPage();
-	} );
+	let readerPage: ReaderPage;
 
 	describe( `Leave a comment as ${ commentingUser }`, function () {
+		let page: Page;
 		let testAccount: TestAccount;
 
-		beforeAll( async () => {
+		beforeAll( async function () {
+			page = await browser.newPage();
 			testAccount = new TestAccount( commentingUser );
 			await testAccount.authenticate( page );
 		} );
 
-		it( 'Visit published site', async function () {
-			// TODO make a utility to obtain a blog URL without string substitution.
-			const siteURL = `https://${ DataHelper.config.get( 'testSiteForNotifications' ) }`;
-			await page.goto( siteURL );
+		it( 'Visit latest post in reader stream', async function () {
+			readerPage = new ReaderPage( page );
+			await readerPage.visitPost( { index: 1 } );
 		} );
 
-		it( 'View first post', async function () {
-			publishedPostsListPage = new PublishedPostsListPage( page );
-			publishedPostsListPage.visitPost( 1 );
-		} );
-
-		it( 'Comment on the post', async function () {
-			const commentsComponent = new CommentsComponent( page );
-			await commentsComponent.postComment( comment );
+		it( 'Leave a comment on the post', async function () {
+			await readerPage.comment( comment );
 		} );
 	} );
 
 	describe( `Trash comment as ${ notificationsUser }`, function () {
-		beforeAll( async () => {
+		let page: Page;
+
+		beforeAll( async function () {
+			page = await browser.newPage();
 			const testAccount = new TestAccount( notificationsUser );
 			await testAccount.authenticate( page );
 		} );
