+++ test/e2e/specs/specs-playwright/wp-reader__view.ts
@@ -1,0 +1,31 @@
+/**
+ * @group calypso-pr
+ */
+
+import { DataHelper, TestAccount, ReaderPage } from '@automattic/calypso-e2e';
+import { Page, Browser } from 'playwright';
+
+declare const browser: Browser;
+
+describe( DataHelper.createSuiteTitle( 'Reader: View' ), function () {
+	let page: Page;
+	let readerPage: ReaderPage;
+	let testAccount: TestAccount;
+
+	beforeAll( async function () {
+		page = await browser.newPage();
+		testAccount = new TestAccount( 'commentingUser' );
+		await testAccount.authenticate( page );
+	} );
+
+	it( 'View the Reader stream', async function () {
+		readerPage = new ReaderPage( page );
+		const testSiteForNotifications = DataHelper.config.get( 'testSiteForNotifications' );
+		const siteOfLatestPost = await readerPage.siteOfLatestPost();
+		expect( siteOfLatestPost ).toEqual( testSiteForNotifications );
+	} );
+
+	it( 'Visit latest post', async function () {
+		await readerPage.visitPost( { index: 1 } );
+	} );
+} );
